requirements.txt
absl-py==2.1.0
altair==5.4.1
astunparse==1.6.3
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==4.4.2
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.66.2
h5py==3.12.1
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
moviepy==1.0.3
namex==0.0.8
narwhals==1.6.4
numba==0.60.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.2
pillow==10.2.0
proglog==0.1.10
protobuf==4.25.5
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytube==15.0.0
pytz==2024.1
referencing==0.35.1
requests==2.31.0
rich==13.8.0
rpds-py==0.20.0
scikit-learn==1.5.1
scipy==1.14.1
setuptools==69.2.0
shap==0.46.0
six==1.16.0
slicer==0.0.8
smmap==5.0.1
streamlit==1.38.0
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.2
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
